Change timeout_ms return type to std::uint32_t
Updated the return type of timeout_ms() from std::uint16_t to std::uint32_t.

+++ main/include/driver/timer/interface.h
@@ -31,7 +31,7 @@ public:
      * @brief Get the current timeout.
      * @return Timeout in milliseconds.
      */
-    virtual std::uint16_t timeout_ms() const noexcept = 0;
+    virtual std::uint32_t timeout_ms() const noexcept = 0;
 
     /**
      * @brief Sets the timer timeout period.
